feat: add clients and refactor structure
