+++ promptfile/types.py
@@ -1,0 +1,6 @@
+from typing import Literal, TypedDict
+
+
+class Message(TypedDict):
+    role: Literal["system", "user", "assistant"]
+    content: str
