+++ setup.py
@@ -10,7 +10,7 @@ with open("README.md", "r") as fh:
 
 setup(
     name="promptfile",
-    version="0.4.0",
+    version="0.5.0",
     packages=find_namespace_packages(),
     entry_points={},
     description="promptfile: language support for .prompt files",
